Feat(crm): Split TasksList component
